added on_save to accelerators
