+++ pytorch_lightning/utilities/cloud_io.py
@@ -55,8 +55,6 @@ def atomic_save(checkpoint, filepath: str):
             This points to the file that the checkpoint will be stored in.
     """
 
-    checkpoint = move_data_to_device(checkpoint, torch.device("cpu"))
-
     bytesbuffer = io.BytesIO()
     # Can't use the new zipfile serialization for 1.6.0 because there's a bug in
     # torch.hub.load_state_dict_from_url() that prevents it from loading the new files.
